No Save Fix on Load
Latest changes ensures that if the save file is empty and we try to load then we will return to previous node if that previous node was a choice. Useful for when you have a menu system which loads a game but the player has not yet started a game. Could potentially disable this option if you wish but for now we simply execute the node again on fail load.

TO DO:

    Need to update docs
    Need to add a save log
    Decide upon allowing players to save and have auto save enabled? - currently both are allowed simultaneously