Radius Increase Causing Completion
Increasing radius (or other methods that may not actually know if the location has been met) will set the failure method to handled when we dont actually know if it is!

The solution was to implement another signal for location services to determine if a location has evaulated to true then find any matching locations in the failure handler and set them to handled!